Doing so allows for splice/sendfile optimizations when available. Since splice can now be leveraged as well, reverse proxy might also benefit tho I haven't confirmed it.
Fixes #4731

To check whether sendfile is really being called or not we can construct two files, any file over 512 bytes should use sendfile in http 1. 
```
$ wc -c *.html
 512 nosendfile.html
 513 sendfile.html
```

logs trimmed for brevity
```
$ sudo strace -f -e sendfile ./caddy file-server --root . --listen 127.0.0.1:8080 --access-log
strace: Process 137302 attached
[...]
2022/09/06 21:56:45.335 INFO    Caddy serving static files on 127.0.0.1:8080
2022/09/06 21:56:50.505 INFO    http.log.access handled request {"request": {"proto": "HTTP/1.1", "method": "GET", "uri": "/nosendfile.html", "size": 512, "status": 200}}
[pid 137301] sendfile(7, 8, NULL, 1)    = 1
2022/09/06 21:57:00.281 INFO    http.log.access handled request {"request": {"proto": "HTTP/1.1", "method": "GET", "uri": "/sendfile.html", "size": 513, "status": 200}}
```

Given this change works under the covers I've added some tests to ensure it won't break.

It would be nice if we could fix `ResponseWriterWrapper` in a less blunt way. Currently, if the underlying writer does not support `ReadFrom` we call `io.Copy` (again). Unfortunately we're fairly constrained on what we can do without breaking everything everywhere so this seemed the most reasonable, backwards compatible, option. The ideal solution would be to not rely on embedding a `ResponseWriterWrapper` but creating one at runtime that implements specifically and exclusively what the underlying writer exposes.

`ResponseRecorder` is now aware of `ReadFrom` too, since it needs to spy on it to gather response size.