@mholt 2022-09-07
Holy shnikies -- I think it works:

```
# (before patch)
$ wrk -t12 -c400 -d10s http://127.0.0.1:1234/index.html
Running 10s test @ http://127.0.0.1:1234/index.html
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    10.32ms   13.39ms 215.51ms   81.90%
    Req/Sec     8.11k     1.10k   19.61k    78.08%
  968720 requests in 10.04s, 35.33GB read
Requests/sec:  96460.94
Transfer/sec:      3.52GB

# (after patch)
$ wrk -t12 -c400 -d10s http://127.0.0.1:1234/index.html
Running 10s test @ http://127.0.0.1:1234/index.html
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.64ms    7.39ms  77.27ms   82.77%
    Req/Sec    12.34k     1.75k   19.47k    68.00%
  1476888 requests in 10.06s, 53.86GB read
Requests/sec: 146868.09
Transfer/sec:      5.36GB
```

Where index.html is just the Caddy homepage, and my Caddyfile is:

```
:1234

root ./test
file_server
```

Bravo, @flga -- you've nearly doubled the performance of the file server over HTTP.

Interestingly, I also see performance improvements with TLS:

```
# (before patch)
$ wrk -t12 -c400 -d10s https://localhost/index.html
Running 10s test @ https://localhost/index.html
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    13.29ms   22.99ms 433.93ms   90.40%
    Req/Sec     6.63k     0.98k    9.87k    74.24%
  769211 requests in 10.09s, 28.08GB read
Requests/sec:  76236.31
Transfer/sec:      2.78GB

# (after patch)
$ wrk -t12 -c400 -d10s https://localhost/index.html
Running 10s test @ https://localhost/index.html
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.19ms   20.59ms 424.09ms   93.98%
    Req/Sec     8.34k     1.58k   17.76k    74.69%
  963389 requests in 10.10s, 35.17GB read
Requests/sec:  95389.21
Transfer/sec:      3.48GB
```

which I would not expect to happen. But I was able to repeat these results. This might be good news for production sites.

I still want to collect profiles during these tests to see where the differences are when sendfile() is used versus when it's not. I might do that later tonight or tomorrow.

@flga Can I invite you to our developers Slack? It's for Caddy maintainers, sponsors, developers, etc. Would love to have you hang out with us there. Just let me know which email to send the invite to.