@francislavoie 2022-09-07
Tried it out for myself as well! I can confirm essentially the same thing:

HTTP Without patch:
```
Requests/sec:  35762.96
Transfer/sec:      1.30GB
Requests/sec:  35393.53
Transfer/sec:      1.29GB
```

HTTP With patch:
```
Requests/sec:  44787.90
Transfer/sec:      1.63GB
Requests/sec:  47676.41
Transfer/sec:      1.74GB
```

HTTPS Without patch:
```
Requests/sec:  29501.75
Transfer/sec:      1.08GB
Requests/sec:  27661.70
Transfer/sec:      1.01GB
```

HTTPS With patch:
```
Requests/sec:  33021.70
Transfer/sec:      1.21GB
Requests/sec:  32188.66
Transfer/sec:      1.17GB
```