@mholt 2022-09-07
CPU profiles were almost identical.

Heap profiles were also pretty similar, but with strikingly fewer allocations and copying across buffers in the patched version (as expected):

![profile001](https://user-images.githubusercontent.com/1128849/188787770-80d37f50-8149-4f2f-bdc5-fea4dd9f9ef7.png)

@flga 

> It would be nice if we could fix ResponseWriterWrapper in a less blunt way. Currently, if the underlying writer does not support ReadFrom we call io.Copy (again). Unfortunately we're fairly constrained on what we can do without breaking everything everywhere so this seemed the most reasonable, backwards compatible, option. The ideal solution would be to not rely on embedding a ResponseWriterWrapper but creating one at runtime that implements specifically and exclusively what the underlying writer exposes.

I agree, I just don't know how to do that... still, I feel like by now we've implemented the most crucial interfaces... I'm open for ideas on better ways to do this though.

Thanks for the test case by the way!

The code change looks good and is just about exactly how I would have done it. Better, even.

I'll do a final review soon and maybe tweak a thing or two, but either way we'll get this merged very shortly. Thank you for this great patch!