@francislavoie 2022-09-07
I tried with 2 Caddy instances, a `file_server` one and a `reverse_proxy` one with the fileserver as an upstream; tried the proxy instance before/after the patch, not seeing any difference (within margin of error).