@mholt 2022-09-07
Verified that this does not affect reverse proxy. (Which is fine, maybe a later patch can address that.)

Our reverse proxy is derived from the std lib's `net/http/httputil` reverse proxy, which implements its own `copyBuffer()` function, which they derived from the `io` package (`io.Copy()`, `io.CopyBuffer()`, etc all use it). Interestingly, the `io.copyBuffer` function _does use_ ReadFrom and WriteTo if available. However, the `httputil.copyBuffer` function does NOT use this, even though much of the remaining code is similar.

Other differences between the two are primarily about error handling.

The omission of this optimization in httputil appears to be intentional: https://github.com/golang/go/issues/21814

However, just for kicks and giggles I added it back into ours to see what the difference is, even though error handling is "incorrect" when the optimization is used. Proxying the same payload as my test above, I got about 35k-36k req/sec. With this patch and re-adding the ReadFrom call, I observed about 40k req/sec. Not a huge boost, but a repeatable result, and still an improvement.

Go's sendfile support only works for TCP connections with an *os.File on one side, even though sendfile(2) supports any file descriptors. That seems unfortunate that we can't copy from one TCP socket to another (unless TCPConns do assert as *os.File? I doubt it?) -- maybe I'm wrong, but that's what it seems like.

So anyway, if this patch were to affect reverse proxy, we'd just need to call ReadFrom in our copyBuffer. It doesn't use sendfile (unless we implement that manually -- which I'd be open to reviewing), but it does reduce allocations/copies. But then, we have the problem of less control over error handling, which is why httputil doesn't just use io.copyBuffer. Hmm.

So I just thought I'd mention this here.

This patch still LGTM!